Move code out of component
